@@ -5,15 +5,15 @@
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <ProjectTypeGuids>{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
     <RootNamespace>Zeiss.PiWeb.Api.Rest</RootNamespace>
-    <TargetFrameworks>netstandard2.0;netcoreapp3.1;net48;net5.0</TargetFrameworks>
+    <TargetFrameworks>netstandard2.0;net6.0</TargetFrameworks>
     <LangVersion>latest</LangVersion>
-    <PackageVersion>7.1.0</PackageVersion>
+    <PackageVersion>8.0.0</PackageVersion>
   </PropertyGroup>
 
   <PropertyGroup Label="Assembly information">
     <AssemblyTitle>A .NET client for HTTP(S)/REST based communication with the quality data managament system ZEISS PiWeb</AssemblyTitle>
     <AssemblyName>Zeiss.PiWeb.Api.Rest</AssemblyName>
-    <AssemblyVersion>7.1.0</AssemblyVersion>
+    <AssemblyVersion>8.0.0</AssemblyVersion>
     <Company>Carl Zeiss Industrielle Messtechnik GmbH</Company>
     <Copyright>Copyright © 2020 $(Company)</Copyright>
     <Description>ZEISS PiWeb-API .NET Client provides an extensive set of methods for reading and writing  inspection plan structure as well
