@@ -15,7 +15,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 	using System;
 	using FluentAssertions;
 	using NUnit.Framework;
-	using Zeiss.PiWeb.Api.Rest.Dtos.Data;
+	using Zeiss.PiWeb.Api.Core;
 
 	#endregion
 
@@ -36,8 +36,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "\u0001";
 			var actualResult = PathHelper.String2PartPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Part( "\u0001" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Part( "\u0001" ) ), actualResult );
 		}
 
 		[Test]
@@ -46,8 +46,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "/foo";
 			var actualResult = PathHelper.String2PartPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Part( "foo" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Part( "foo" ) ), actualResult );
 		}
 
 		[Test]
@@ -56,8 +56,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "/\u0001";
 			var actualResult = PathHelper.String2PartPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Part( "\u0001" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Part( "\u0001" ) ), actualResult );
 		}
 
 		[Test]
@@ -66,8 +66,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "foo/";
 			var actualResult = PathHelper.String2PartPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Part( "foo" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Part( "foo" ) ), actualResult );
 		}
 
 		[Test]
@@ -76,8 +76,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "\u0001/";
 			var actualResult = PathHelper.String2PartPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Part( "\u0001" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Part( "\u0001" ) ), actualResult );
 		}
 
 		[Test]
@@ -86,8 +86,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "\\foo";
 			var actualResult = PathHelper.String2PartPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Part( "foo" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Part( "foo" ) ), actualResult );
 		}
 
 		[Test]
@@ -96,8 +96,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "/";
 			var actualResult = PathHelper.String2PartPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( PathInformationDto.Root, actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( PathInformation.Root, actualResult );
 		}
 
 		[Test]
@@ -106,8 +106,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "/foo/bar/";
 			var actualResult = PathHelper.String2PartPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Part( "foo" ), PathElementDto.Part( "bar" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Part( "foo" ), PathElement.Part( "bar" ) ), actualResult );
 		}
 
 		[Test]
@@ -138,8 +138,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "\u0001";
 			var actualResult = PathHelper.String2CharPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Char( "\u0001" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Char( "\u0001" ) ), actualResult );
 		}
 
 		[Test]
@@ -148,8 +148,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "/foo";
 			var actualResult = PathHelper.String2CharPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Char( "foo" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Char( "foo" ) ), actualResult );
 		}
 
 		[Test]
@@ -158,8 +158,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "/\u0001";
 			var actualResult = PathHelper.String2CharPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Char( "\u0001" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Char( "\u0001" ) ), actualResult );
 		}
 
 		[Test]
@@ -168,8 +168,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "foo/";
 			var actualResult = PathHelper.String2CharPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Char( "foo" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Char( "foo" ) ), actualResult );
 		}
 
 		[Test]
@@ -178,8 +178,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "\u0001/";
 			var actualResult = PathHelper.String2CharPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Char( "\u0001" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Char( "\u0001" ) ), actualResult );
 		}
 
 		[Test]
@@ -196,8 +196,8 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "/foo/bar/";
 			var actualResult = PathHelper.String2CharPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Char( "foo" ), PathElementDto.Char( "bar" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Char( "foo" ), PathElement.Char( "bar" ) ), actualResult );
 		}
 
 		[Test]
@@ -222,8 +222,192 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Tests
 			const string path = "\\foo";
 			var actualResult = PathHelper.String2CharPathInformation( path );
 
-			actualResult.Should().BeOfType<PathInformationDto>();
-			Assert.AreEqual( new PathInformationDto( PathElementDto.Char( "foo" ) ), actualResult );
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Char( "foo" ) ), actualResult );
+		}
+
+		[Test]
+		public void String2CharPathInformation_PathNotEndingWithUnquotedDelimiter_ThrowsCorrectException()
+		{
+			const string path = "\\foo\\";
+
+			FluentActions.Invoking( () => PathHelper.String2CharPathInformation( path ) ).Should().Throw<InvalidOperationException>()
+				.Where( e => e.Message.Contains( "does not end with an unquoted delimiter character" ) );
+		}
+
+		[Test]
+		public void RoundtripString2PathInformation_PathIsNull_ThrowsCorrectException()
+		{
+			Assert.Throws<ArgumentException>( () => PathHelper.RoundtripString2PathInformation( null! ) );
+		}
+
+		[Test]
+		public void RoundtripString2PathInformation_PathIsEmpty_ThrowsCorrectException()
+		{
+			Assert.Throws<ArgumentException>( () => PathHelper.RoundtripString2PathInformation( null! ) );
+		}
+
+		[Test]
+		public void RoundtripString2PathInformation_HappyPath_ReturnsPathInformationDto()
+		{
+			const string path = "PP:/foo/bar/";
+			var actualResult = PathHelper.RoundtripString2PathInformation( path );
+
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Part( "foo" ), PathElement.Part( "bar" ) ), actualResult );
+		}
+
+		[Test]
+		public void RoundtripString2PathInformation_PartsAndCharacteristic_ReturnsPathInformationDto()
+		{
+			const string path = "PPC:/foo/bar/char1/";
+			var actualResult = PathHelper.RoundtripString2PathInformation( path );
+
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( new PathInformation( PathElement.Part( "foo" ), PathElement.Part( "bar" ), PathElement.Char( "char1" ) ), actualResult );
+		}
+
+		[Test]
+		public void RoundtripString2PathInformation_MissingStructure_ThrowsCorrectException()
+		{
+			const string path = "/foo/bar/char1/";
+
+			FluentActions.Invoking( () => PathHelper.RoundtripString2PathInformation( path ) ).Should().Throw<ArgumentException>()
+				.WithMessage( "The path must have the following structure:\"structure:path\", e.g.: \"PC:/part/characteristic/\"." );
+		}
+
+		[Test]
+		public void RoundtripString2PathInformation_WrongStructure_ThrowsCorrectException()
+		{
+			const string path = "PP:/foo/bar/char1/";
+
+			FluentActions.Invoking( () => PathHelper.RoundtripString2PathInformation( path ) ).Should().Throw<ArgumentException>();
+		}
+
+		[Test]
+		public void RoundtripString2PathInformation_FastPathRoot_ReturnsPathInformationDto()
+		{
+			const string path = "/";
+			var actualResult = PathHelper.RoundtripString2PathInformation( path );
+
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( PathInformation.Root, actualResult );
+		}
+
+		[Test]
+		public void RoundtripString2PathInformation_MissingFirstDelimiter_ThrowsException()
+		{
+			const string path = "P:NoDelimiter/";
+
+			FluentActions.Invoking( () => PathHelper.RoundtripString2PathInformation( path ) ).Should().Throw<InvalidOperationException>()
+				.Where( e => e.Message.Contains( "The first character of path string" ) );
+		}
+
+		[Test]
+		public void DatabaseString2PathInformation_PathIsNull_ThrowsCorrectException()
+		{
+			Assert.Throws<ArgumentException>( () => PathHelper.DatabaseString2PathInformation( null!, "".AsSpan() ) );
+		}
+
+		[Test]
+		public void DatabaseString2PathInformation_FastPathRoot_ReturnsPathInformationDto()
+		{
+			const string path = "/";
+			var actualResult = PathHelper.DatabaseString2PathInformation( path.AsSpan(), "".AsSpan() );
+
+			actualResult.Should().BeOfType<PathInformation>();
+			Assert.AreEqual( PathInformation.Root, actualResult );
+		}
+
+		[Test]
+		public void DatabaseString2PathInformation_MissingLastDelimiter_ThrowsCorrectException()
+		{
+			const string path = "/foo/bar";
+
+			FluentActions.Invoking( () => PathHelper.DatabaseString2PathInformation( path.AsSpan(), "PP".AsSpan() ) ).Should().Throw<InvalidOperationException>()
+				.Where( e => e.Message.Contains( "The last character of path string" ) );
+		}
+
+		[Test]
+		public void PathInformation2String_PathIsNull_ThrowsCorrectException()
+		{
+			Assert.Throws<ArgumentNullException>( () => PathHelper.PathInformation2String( null! ) );
+		}
+
+		[Test]
+		public void PathInformation2String_HappyPath_ReturnsCorrectString()
+		{
+			var pathInformation = new PathInformation( PathElement.Part( "foo" ), PathElement.Part( "bar" ) );
+			var actualResult = PathHelper.PathInformation2String( pathInformation );
+
+			const string expected = "foo/bar";
+
+			Assert.AreEqual( expected, actualResult );
+		}
+
+		[Test]
+		public void PathInformation2String_FastPathRoot_ReturnsDelimiter()
+		{
+			var actualResult = PathHelper.PathInformation2String( PathInformation.Root );
+
+			Assert.AreEqual( PathHelper.DelimiterString, actualResult );
+		}
+
+		[Test]
+		public void PathInformation2RoundtripString_PathIsNull_ThrowsCorrectException()
+		{
+			Assert.Throws<ArgumentNullException>( () => PathHelper.PathInformation2RoundtripString( null! ) );
+		}
+
+		[Test]
+		public void PathInformation2RoundtripString_HappyPath_ReturnsCorrectString()
+		{
+			var pathInformation = new PathInformation( PathElement.Part( "foo" ), PathElement.Part( "bar" ), PathElement.Char( "char1" ) );
+			var actualResult = PathHelper.PathInformation2RoundtripString( pathInformation );
+
+			const string expected = "PPC:/foo/bar/char1/";
+
+			Assert.AreEqual( expected, actualResult );
+		}
+
+		[Test]
+		public void PathInformation2RoundtripString_FastPathRoot_ReturnsDelimiter()
+		{
+			var actualResult = PathHelper.PathInformation2RoundtripString( PathInformation.Root );
+
+			Assert.AreEqual( PathHelper.DelimiterString, actualResult );
+		}
+
+		[Test]
+		public void PathInformation2DatabaseString_PathIsNull_ThrowsCorrectException()
+		{
+			Assert.Throws<ArgumentNullException>( () => PathHelper.PathInformation2DatabaseString( null! ) );
+		}
+
+
+		[Test]
+		public void PathInformation2DatabaseString_FastPathRoot_ReturnsDelimiter()
+		{
+			var actualResult = PathHelper.PathInformation2DatabaseString( PathInformation.Root );
+
+			Assert.AreEqual( PathHelper.DelimiterString, actualResult );
+		}
+
+		[Test]
+		public void GetStructure_PathIsNull_ThrowsCorrectException()
+		{
+			Assert.Throws<ArgumentNullException>( () => PathHelper.GetStructure( null! ) );
+		}
+
+		[Test]
+		public void GetStructure_HappyPath_ReturnsCorrectString()
+		{
+			var pathInformation = new PathInformation( PathElement.Part( "foo" ), PathElement.Part( "bar" ), PathElement.Char( "char1" ) );
+			var actualResult = PathHelper.GetStructure( pathInformation );
+
+			const string expected = "PPC";
+
+			Assert.AreEqual( expected, actualResult );
 		}
 
 		#endregion
