@@ -13,9 +13,10 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	#region usings
 
 	using System;
+	using System.Collections.Generic;
+	using System.Text.Json.Serialization;
 	using JetBrains.Annotations;
 	using Newtonsoft.Json;
-	using Zeiss.PiWeb.Api.Rest.Dtos.Converter;
 
 	#endregion
 
@@ -26,7 +27,9 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	{
 		#region members
 
-		private DataCharacteristicDto[] _Characteristics = Array.Empty<DataCharacteristicDto>();
+		// Newtonsoft.Json is not able to create an empty dictionary if the property is not set.
+		// If Newtonsoft.Json support is discontinued, this initialization should no longer be necessary.
+		[NotNull] private IReadOnlyDictionary<Guid, DataValueDto> _Characteristics = new Dictionary<Guid, DataValueDto>();
 
 		#endregion
 
@@ -35,11 +38,12 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 		/// <summary>
 		/// Gets or sets the measurement values.
 		/// </summary>
-		[JsonProperty( "characteristics" ), JsonConverter( typeof( DataCharacteristicConverter ) )]
-		public DataCharacteristicDto[] Characteristics
+		[JsonProperty( "characteristics" )]
+		[JsonPropertyName( "characteristics" )]
+		public IReadOnlyDictionary<Guid, DataValueDto> Characteristics
 		{
 			[NotNull] get => _Characteristics;
-			set => _Characteristics = value ?? Array.Empty<DataCharacteristicDto>();
+			set => _Characteristics = value ?? new Dictionary<Guid, DataValueDto>();
 		}
 
 		#endregion
