@@ -21,7 +21,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	#endregion
 
 	/// <summary>
-	/// Class that encapsulates the url parameter for a measurement value search via PiWeb-REST web service.
+	/// This class contains a measurement search criteria for a measurement search.
 	/// </summary>
 	public class MeasurementFilterAttributesDto : AbstractMeasurementFilterAttributesDto
 	{
@@ -68,7 +68,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 		/// <summary>
 		/// Specifies the list of primary measurement keys to be used for joining measurements accross multiple parts on the server side.
 		/// </summary>
-		public ushort[] MergeAttributes { get; set; }
+		public IReadOnlyList<ushort> MergeAttributes { get; set; }
 
 		/// <summary>
 		/// Specifies the condition that must be adhered to
@@ -112,21 +112,21 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 		{
 			var items = new[]
 			{
-				Tuple.Create( PartUuidsParamName, partUuids ),
-				Tuple.Create( MeasurementUuidsParamName, measurementUuids ),
-				Tuple.Create( DeepParamName, deep ),
-				Tuple.Create( LimitResultParamName, limitResult ),
-				Tuple.Create( LimitResultPerPartParamName, limitResultPerPart ),
-				Tuple.Create( OrderByParamName, order ),
-				Tuple.Create( RequestedMeasurementAttributesParamName, requestedMeasurementAttributes ),
-				Tuple.Create( SearchConditionParamName, searchCondition ),
-				Tuple.Create( StatisticsParamName, statistics ),
-				Tuple.Create( AggregationParamName, aggregation ),
-				Tuple.Create( FromModificationDateParamName, fromModificationDate ),
-				Tuple.Create( ToModificationDateParamName, toModificationDate ),
-				Tuple.Create( MergeAttributesParamName, mergeAttributes ),
-				Tuple.Create( MergeConditionParamName, mergeCondition ),
-				Tuple.Create( MergeMasterPartParamName, mergeMasterPart )
+				ValueTuple.Create( PartUuidsParamName, partUuids ),
+				ValueTuple.Create( MeasurementUuidsParamName, measurementUuids ),
+				ValueTuple.Create( DeepParamName, deep ),
+				ValueTuple.Create( LimitResultParamName, limitResult ),
+				ValueTuple.Create( LimitResultPerPartParamName, limitResultPerPart ),
+				ValueTuple.Create( OrderByParamName, order ),
+				ValueTuple.Create( RequestedMeasurementAttributesParamName, requestedMeasurementAttributes ),
+				ValueTuple.Create( SearchConditionParamName, searchCondition ),
+				ValueTuple.Create( StatisticsParamName, statistics ),
+				ValueTuple.Create( AggregationParamName, aggregation ),
+				ValueTuple.Create( FromModificationDateParamName, fromModificationDate ),
+				ValueTuple.Create( ToModificationDateParamName, toModificationDate ),
+				ValueTuple.Create( MergeAttributesParamName, mergeAttributes ),
+				ValueTuple.Create( MergeConditionParamName, mergeCondition ),
+				ValueTuple.Create( MergeMasterPartParamName, mergeMasterPart )
 			};
 
 			var result = new MeasurementFilterAttributesDto();
@@ -247,11 +247,11 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 		}
 
 		/// <inheritdoc />
-		public override ParameterDefinition[] ToParameterDefinition()
+		public override IReadOnlyCollection<ParameterDefinition> ToParameterDefinition()
 		{
 			var result = new List<ParameterDefinition>();
 
-			if( PartUuids != null && PartUuids.Length > 0 )
+			if( PartUuids != null && PartUuids.Count > 0 )
 				result.Add( ParameterDefinition.Create( PartUuidsParamName, RestClientHelper.ConvertGuidListToString( PartUuids ) ) );
 
 			if( Deep )
@@ -263,13 +263,13 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 			if( LimitResultPerPart >= 0 )
 				result.Add( ParameterDefinition.Create( LimitResultPerPartParamName, LimitResultPerPart.ToString() ) );
 
-			if( MeasurementUuids != null && MeasurementUuids.Length > 0 )
+			if( MeasurementUuids != null && MeasurementUuids.Count > 0 )
 				result.Add( ParameterDefinition.Create( MeasurementUuidsParamName, RestClientHelper.ConvertGuidListToString( MeasurementUuids ) ) );
 
 			if( RequestedMeasurementAttributes != null && RequestedMeasurementAttributes.AllAttributes != AllAttributeSelectionDto.True && RequestedMeasurementAttributes.Attributes != null )
 				result.Add( ParameterDefinition.Create( RequestedMeasurementAttributesParamName, RestClientHelper.ConvertUshortArrayToString( RequestedMeasurementAttributes.Attributes ) ) );
 
-			if( OrderBy != null && OrderBy.Length > 0 )
+			if( OrderBy != null && OrderBy.Count > 0 )
 				result.Add( ParameterDefinition.Create( OrderByParamName, OrderByToString( OrderBy ) ) );
 
 			if( SearchCondition != null )
@@ -287,7 +287,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 			if( ToModificationDate.HasValue )
 				result.Add( ParameterDefinition.Create( ToModificationDateParamName, XmlConvert.ToString( ToModificationDate.Value, XmlDateTimeSerializationMode.RoundtripKind ) ) );
 
-			if( MergeAttributes != null && MergeAttributes.Length > 0 )
+			if( MergeAttributes != null && MergeAttributes.Count > 0 )
 				result.Add( ParameterDefinition.Create( MergeAttributesParamName, RestClientHelper.ConvertUshortArrayToString( MergeAttributes ) ) );
 
 			if( MergeCondition != MeasurementMergeConditionDto.MeasurementsInAllParts )
@@ -296,7 +296,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 			if( MergeMasterPart != null )
 				result.Add( ParameterDefinition.Create( MergeMasterPartParamName, MergeMasterPart.ToString() ) );
 
-			return result.ToArray();
+			return result;
 		}
 
 		#endregion
