@@ -19,6 +19,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 	using System.Threading;
 	using System.Threading.Tasks;
 	using JetBrains.Annotations;
+	using Zeiss.PiWeb.Api.Core;
 	using Zeiss.PiWeb.Api.Definitions;
 	using Zeiss.PiWeb.Api.Rest.Common.Client;
 	using Zeiss.PiWeb.Api.Rest.Common.Data;
@@ -55,7 +56,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		/// <param name="maxUriLength">The uri length limit</param>
 		/// <param name="restClient">Custom implementation of RestClient</param>
 		public DataServiceRestClient( [NotNull] Uri serverUri, int maxUriLength = RestClientBase.DefaultMaxUriLength, RestClientBase restClient = null )
-			: base( restClient ?? new RestClient( serverUri, EndpointName, maxUriLength: maxUriLength ) )
+			: base( restClient ?? new RestClient( serverUri, EndpointName, maxUriLength: maxUriLength, serializer: ObjectSerializer.SystemTextJson ) )
 		{ }
 
 		#endregion
@@ -87,7 +88,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 			}
 		}
 
-		private async Task<DataMeasurementDto[]> GetMeasurementValuesSplitByMeasurement( PathInformationDto partPath, MeasurementValueFilterAttributesDto filter, CancellationToken cancellationToken )
+		private async Task<IReadOnlyList<DataMeasurementDto>> GetMeasurementValuesSplitByMeasurement( PathInformation partPath, MeasurementValueFilterAttributesDto filter, CancellationToken cancellationToken )
 		{
 			var newFilter = filter.Clone();
 			newFilter.MeasurementUuids = null;
@@ -98,27 +99,27 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 			var requestRestriction = RequestBuilder.AppendParameters( "values", parameter );
 			var targetSize = RestClientHelper.GetUriTargetSize( ServiceLocation, requestRestriction, MaxUriLength );
 
-			var result = new List<DataMeasurementDto>( filter.MeasurementUuids.Length );
+			var result = new List<DataMeasurementDto>( filter.MeasurementUuids.Count );
 			var resultSets = 0;
 			foreach( var uuids in ArrayHelper.Split( filter.MeasurementUuids, targetSize, RestClientHelper.LengthOfListElementInUri ) )
 			{
 				newFilter.MeasurementUuids = uuids;
-				if( newFilter.CharacteristicsUuidList?.Length > 0 )
+				if( newFilter.CharacteristicsUuidList?.Count > 0 )
 				{
 					result.AddRange( await GetMeasurementValuesSplitByCharacteristics( partPath, newFilter, cancellationToken ).ConfigureAwait( false ) );
 				}
 				else
 				{
-					result.AddRange( await _RestClient.Request<DataMeasurementDto[]>( RequestBuilder.CreateGet( "values", CreateParameterDefinitions( partPath, newFilter ).ToArray() ), cancellationToken ).ConfigureAwait( false ) );
+					result.AddRange( await _RestClient.Request<IReadOnlyList<DataMeasurementDto>>( RequestBuilder.CreateGet( "values", CreateParameterDefinitions( partPath, newFilter ) ), cancellationToken ).ConfigureAwait( false ) );
 				}
 
 				resultSets++;
 			}
 
-			return resultSets > 1 ? LimitAndSortResult( result, filter, resultSets ).Cast<DataMeasurementDto>().ToArray() : result.ToArray();
+			return resultSets > 1 ? LimitAndSortResult( result, filter, resultSets ) : result;
 		}
 
-		private async Task<DataMeasurementDto[]> GetMeasurementValuesSplitByCharacteristics( PathInformationDto partPath, MeasurementValueFilterAttributesDto filter, CancellationToken cancellationToken )
+		private async Task<IReadOnlyList<DataMeasurementDto>> GetMeasurementValuesSplitByCharacteristics( PathInformation partPath, MeasurementValueFilterAttributesDto filter, CancellationToken cancellationToken )
 		{
 			var newFilter = filter.Clone();
 			newFilter.CharacteristicsUuidList = null;
@@ -129,13 +130,13 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 			var requestRestriction = RequestBuilder.AppendParameters( "values", parameter );
 			var targetSize = RestClientHelper.GetUriTargetSize( ServiceLocation, requestRestriction, MaxUriLength );
 
-			var result = new List<DataMeasurementDto>( filter.MeasurementUuids?.Length ?? 0 );
+			var result = new List<DataMeasurementDto>( filter.MeasurementUuids?.Count ?? 0 );
 			var allMeasurements = new Dictionary<Guid, DataMeasurementDto>();
 			foreach( var uuids in ArrayHelper.Split( filter.CharacteristicsUuidList, targetSize, RestClientHelper.LengthOfListElementInUri ) )
 			{
 				newFilter.CharacteristicsUuidList = uuids;
 
-				var measurements = await _RestClient.Request<DataMeasurementDto[]>( RequestBuilder.CreateGet( "values", CreateParameterDefinitions( partPath, newFilter ).ToArray() ), cancellationToken ).ConfigureAwait( false );
+				var measurements = await _RestClient.Request<IEnumerable<DataMeasurementDto>>( RequestBuilder.CreateGet( "values", CreateParameterDefinitions( partPath, newFilter ) ), cancellationToken ).ConfigureAwait( false );
 				foreach( var measurement in measurements )
 				{
 					if( allMeasurements.TryGetValue( measurement.Uuid, out var existingMeasurement ) )
@@ -150,20 +151,30 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 				}
 			}
 
-			return result.ToArray();
+			return result;
 		}
 
-		private static DataCharacteristicDto[] Combine( DataCharacteristicDto[] list1, DataCharacteristicDto[] list2 )
+		private static IReadOnlyDictionary<Guid, DataValueDto> Combine( IReadOnlyDictionary<Guid, DataValueDto> list1, IReadOnlyDictionary<Guid, DataValueDto> list2 )
 		{
 			if( list1 == null )
 				return list2;
 			if( list2 == null )
 				return list1;
 
-			return list1.Concat( list2 ).ToArray();
+			var result = new Dictionary<Guid, DataValueDto>( list1.Count + list2.Count );
+			foreach( var item in list1 )
+			{
+				result.Add( item.Key, item.Value );
+			}
+			foreach( var item in list2 )
+			{
+				result.Add( item.Key, item.Value );
+			}
+
+			return result;
 		}
 
-		private static List<ParameterDefinition> CreateParameterDefinitions<T>( PathInformationDto partPath, T filter, int? key = null ) where T : AbstractMeasurementFilterAttributesDto
+		private static List<ParameterDefinition> CreateParameterDefinitions<T>( PathInformation partPath, T filter, int? key = null ) where T : AbstractMeasurementFilterAttributesDto
 		{
 			var parameter = new List<ParameterDefinition>();
 
@@ -214,12 +225,17 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		/// <param name="filter">Set filter to check for LimitResult and OrderBy</param>
 		/// <param name="resultSets">Number of requests after splitting the initial request</param>
 		/// <returns>Array of measurements in correct order and size.</returns>
-		private static IEnumerable<SimpleMeasurementDto> LimitAndSortResult( IEnumerable<SimpleMeasurementDto> source, AbstractMeasurementFilterAttributesDto filter, int resultSets )
+		private static IReadOnlyList<T> LimitAndSortResult<T>( IEnumerable<T> source, AbstractMeasurementFilterAttributesDto filter, int resultSets ) where T: SimpleMeasurementDto
 		{
-			if( resultSets <= 1 ) return source;
+			if( resultSets <= 1 )
+				return ConvertToList( source );
 
 			if( filter.OrderBy == null || !filter.OrderBy.Any() )
-				return filter.LimitResult < 0 ? source : source.Take( filter.LimitResult );
+			{
+				return filter.LimitResult < 0
+					? ConvertToList( source )
+					: source.Take( filter.LimitResult ).ToArray();
+			}
 
 			var orderedResult = StartAttributeOrderChain( filter.OrderBy.First(), source );
 
@@ -229,21 +245,28 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 				orderedResult = AppendAttributeOrderChain( orderDto, orderedResult );
 			}
 
-			if( orderedResult == null )
-				return Array.Empty<SimpleMeasurementDto>();
+			return filter.LimitResult < 0
+				? orderedResult.ToArray()
+				: orderedResult.Take( filter.LimitResult ).ToArray();
+		}
 
-			return filter.LimitResult < 0 ? orderedResult : orderedResult.Take( filter.LimitResult );
+		private static IReadOnlyList<T> ConvertToList<T>( IEnumerable<T> source ) where T : SimpleMeasurementDto
+		{
+			if( source is IReadOnlyList<T> list )
+				return list;
+
+			return source.ToArray();
 		}
 
 		/// <summary>
 		/// Sort the measurements according to the order criteria. Start a new OrderBy-chain.
 		/// </summary>
-		private static IOrderedEnumerable<SimpleMeasurementDto> StartAttributeOrderChain( OrderDto order, IEnumerable<SimpleMeasurementDto> source )
+		private static IOrderedEnumerable<T> StartAttributeOrderChain<T>( OrderDto order, IEnumerable<T> source ) where T: SimpleMeasurementDto
 		{
 			return order.Direction switch
 			{
-				OrderDirectionDto.Asc => source.OrderBy( SelectAttributeValues( order ) ),
-				OrderDirectionDto.Desc => source.OrderByDescending( SelectAttributeValues( order ) ),
+				OrderDirectionDto.Asc => source.OrderBy( SelectAttributeValues<T>( order ) ),
+				OrderDirectionDto.Desc => source.OrderByDescending( SelectAttributeValues<T>( order ) ),
 				_ => throw new ArgumentOutOfRangeException( nameof( order.Direction ) )
 			};
 		}
@@ -251,24 +274,22 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		/// <summary>
 		/// Sort the measurements according to the order criteria. Chain it to a previous result.
 		/// </summary>
-		private static IOrderedEnumerable<SimpleMeasurementDto> AppendAttributeOrderChain( OrderDto order, IOrderedEnumerable<SimpleMeasurementDto> orderedSource )
+		private static IOrderedEnumerable<T> AppendAttributeOrderChain<T>( OrderDto order, IOrderedEnumerable<T> orderedSource ) where T: SimpleMeasurementDto
 		{
 			return order.Direction switch
 			{
-				OrderDirectionDto.Asc => orderedSource.ThenBy( SelectAttributeValues( order ) ),
-				OrderDirectionDto.Desc => orderedSource.ThenByDescending( SelectAttributeValues( order ) ),
+				OrderDirectionDto.Asc => orderedSource.ThenBy( SelectAttributeValues<T>( order ) ),
+				OrderDirectionDto.Desc => orderedSource.ThenByDescending( SelectAttributeValues<T>( order ) ),
 				_ => throw new ArgumentOutOfRangeException( nameof( order.Direction ) )
 			};
 		}
 
-		private static Func<SimpleMeasurementDto, object> SelectAttributeValues( OrderDto order )
+		private static Func<T, object> SelectAttributeValues<T>( OrderDto order ) where  T: SimpleMeasurementDto
 		{
 			if( order.Attribute == WellKnownKeys.Measurement.Time )
-			{
-				return m => m.Time;
-			}
+				return measurement => measurement.Time;
 
-			return m => m.Attributes.FirstOrDefault( a => a.Key == order.Attribute )?.Value;
+			return measurement => measurement.GetAttributeValue( order.Attribute );
 		}
 
 		private async Task PollOperationStatus( string statusId, CancellationToken cancellationToken = default )
@@ -290,7 +311,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 			if( featureMatrix.SupportsAsyncMeasurementDeletion )
 			{
 				parameter.Add( ParameterDefinition.Create( "runAsync", "true" ) );
-				var pollingUri = await _RestClient.RequestAsyncOperation( RequestBuilder.CreateDelete( "measurements", parameter.ToArray() ), cancellationToken ).ConfigureAwait( false );
+				var pollingUri = await _RestClient.RequestAsyncOperation( RequestBuilder.CreateDelete( "measurements", parameter ), cancellationToken ).ConfigureAwait( false );
 
 				if( pollingUri != null )
 				{
@@ -300,7 +321,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 			}
 			else
 			{
-				await _RestClient.Request( RequestBuilder.CreateDelete( "measurements", parameter.ToArray() ), cancellationToken ).ConfigureAwait( false );
+				await _RestClient.Request( RequestBuilder.CreateDelete( "measurements", parameter ), cancellationToken ).ConfigureAwait( false );
 			}
 		}
 
@@ -440,7 +461,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="definitions"/> is <see langword="null"/>.</exception>
-		public Task CreateAttributeDefinitions( EntityDto entity, AbstractAttributeDefinitionDto[] definitions, CancellationToken cancellationToken = default )
+		public Task CreateAttributeDefinitions( EntityDto entity, IReadOnlyCollection<AbstractAttributeDefinitionDto> definitions, CancellationToken cancellationToken = default )
 		{
 			if( definitions == null ) throw new ArgumentNullException( nameof( definitions ) );
 			return _RestClient.Request( RequestBuilder.CreatePost( $"configuration/{entity}", Payload.Create( definitions ) ), cancellationToken );
@@ -448,18 +469,18 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="definitions"/> is <see langword="null"/>.</exception>
-		public Task UpdateAttributeDefinitions( EntityDto entity, AbstractAttributeDefinitionDto[] definitions, CancellationToken cancellationToken = default )
+		public Task UpdateAttributeDefinitions( EntityDto entity, IReadOnlyCollection<AbstractAttributeDefinitionDto> definitions, CancellationToken cancellationToken = default )
 		{
 			if( definitions == null ) throw new ArgumentNullException( nameof( definitions ) );
 			return _RestClient.Request( RequestBuilder.CreatePut( $"configuration/{entity}", Payload.Create( definitions ) ), cancellationToken );
 		}
 
 		/// <inheritdoc />
-		public async Task DeleteAttributeDefinitions( EntityDto entity, ushort[] keys = null, CancellationToken cancellationToken = default )
+		public async Task DeleteAttributeDefinitions( EntityDto entity, IReadOnlyCollection<ushort> keys = null, CancellationToken cancellationToken = default )
 		{
 			var requestBasePath = $"configuration/{entity}";
 
-			if( keys == null || keys.Length == 0 )
+			if( keys == null || keys.Count == 0 )
 			{
 				await _RestClient.Request( RequestBuilder.CreateDelete( requestBasePath ), cancellationToken ).ConfigureAwait( false );
 			}
@@ -487,30 +508,30 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		}
 
 		/// <inheritdoc />
-		public Task<bool> CheckCatalogEntryUsage( ushort attributeKey, int catalogEntryIndex, CancellationToken cancellationToken = default )
+		public Task<bool> CheckCatalogEntryUsage( ushort attributeKey, short catalogEntryKey, CancellationToken cancellationToken = default )
 		{
-			if( catalogEntryIndex < 0 )
-				throw new InvalidOperationException( $"Unable to check catalogue entry usage. {nameof( catalogEntryIndex )} must be equal or greater than 0." );
+			if( catalogEntryKey < 0 )
+				throw new InvalidOperationException( $"Unable to check catalogue entry usage. {nameof( catalogEntryKey )} must be equal or greater than 0." );
 
-			return CheckAttributeUsageInternal( attributeKey, catalogEntryIndex.ToString(), cancellationToken );
+			return CheckAttributeUsageInternal( attributeKey, catalogEntryKey.ToString(), cancellationToken );
 		}
 
 		/// <inheritdoc />
-		public Task<CatalogDto[]> GetAllCatalogs( CancellationToken cancellationToken = default )
+		public Task<IReadOnlyList<CatalogDto>> GetAllCatalogs( CancellationToken cancellationToken = default )
 		{
-			return _RestClient.Request<CatalogDto[]>( RequestBuilder.CreateGet( "catalogs" ), cancellationToken );
+			return _RestClient.Request<IReadOnlyList<CatalogDto>>( RequestBuilder.CreateGet( "catalogs" ), cancellationToken );
 		}
 
 		/// <inheritdoc />
 		public async Task<CatalogDto> GetCatalog( Guid catalogUuid, CancellationToken cancellationToken = default )
 		{
-			var catalog = await _RestClient.Request<CatalogDto[]>( RequestBuilder.CreateGet( $"catalogs/{catalogUuid}" ), cancellationToken ).ConfigureAwait( false );
+			var catalog = await _RestClient.Request<IEnumerable<CatalogDto>>( RequestBuilder.CreateGet( $"catalogs/{catalogUuid}" ), cancellationToken ).ConfigureAwait( false );
 			return catalog.FirstOrDefault();
 		}
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="catalogs"/> is <see langword="null"/>.</exception>
-		public Task CreateCatalogs( CatalogDto[] catalogs, CancellationToken cancellationToken = default )
+		public Task CreateCatalogs( IReadOnlyCollection<CatalogDto> catalogs, CancellationToken cancellationToken = default )
 		{
 			if( catalogs == null ) throw new ArgumentNullException( nameof( catalogs ) );
 			return _RestClient.Request( RequestBuilder.CreatePost( "catalogs", Payload.Create( catalogs ) ), cancellationToken );
@@ -518,18 +539,18 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="catalogs"/> is <see langword="null"/>.</exception>
-		public Task UpdateCatalogs( CatalogDto[] catalogs, CancellationToken cancellationToken = default )
+		public Task UpdateCatalogs( IReadOnlyCollection<CatalogDto> catalogs, CancellationToken cancellationToken = default )
 		{
 			if( catalogs == null ) throw new ArgumentNullException( nameof( catalogs ) );
 			return _RestClient.Request( RequestBuilder.CreatePut( "catalogs", Payload.Create( catalogs ) ), cancellationToken );
 		}
 
 		/// <inheritdoc />
-		public async Task DeleteCatalogs( Guid[] catalogUuids = null, CancellationToken cancellationToken = default )
+		public async Task DeleteCatalogs( IReadOnlyCollection<Guid> catalogUuids = null, CancellationToken cancellationToken = default )
 		{
 			const string uri = "catalogs";
 
-			if( catalogUuids == null || catalogUuids.Length == 0 )
+			if( catalogUuids == null || catalogUuids.Count == 0 )
 			{
 				await _RestClient.Request( RequestBuilder.CreateDelete( uri ), cancellationToken ).ConfigureAwait( false );
 			}
@@ -546,18 +567,18 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="entries"/> is <see langword="null"/>.</exception>
-		public Task CreateCatalogEntries( Guid catalogUuid, CatalogEntryDto[] entries, CancellationToken cancellationToken = default )
+		public Task CreateCatalogEntries( Guid catalogUuid, IReadOnlyCollection<CatalogEntryDto> entries, CancellationToken cancellationToken = default )
 		{
 			if( entries == null ) throw new ArgumentNullException( nameof( entries ) );
 			return _RestClient.Request( RequestBuilder.CreatePost( $"catalogs/{catalogUuid}", Payload.Create( entries ) ), cancellationToken );
 		}
 
 		/// <inheritdoc />
-		public async Task DeleteCatalogEntries( Guid catalogUuid, short[] keys = null, CancellationToken cancellationToken = default )
+		public async Task DeleteCatalogEntries( Guid catalogUuid, IReadOnlyCollection<short> keys = null, CancellationToken cancellationToken = default )
 		{
 			var uri = $"catalogs/{catalogUuid}";
 
-			if( keys == null || keys.Length == 0 )
+			if( keys == null || keys.Count == 0 )
 			{
 				await _RestClient.Request( RequestBuilder.CreateDelete( uri ), cancellationToken ).ConfigureAwait( false );
 			}
@@ -572,11 +593,11 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		}
 
 		/// <inheritdoc />
-		public async Task<IEnumerable<InspectionPlanPartDto>> GetParts( PathInformationDto partPath = null, Guid[] partUuids = null, ushort? depth = null, AttributeSelector requestedPartAttributes = null, bool withHistory = false, CancellationToken cancellationToken = default )
+		public async Task<IReadOnlyList<InspectionPlanPartDto>> GetParts( PathInformation partPath = null, IReadOnlyCollection<Guid> partUuids = null, ushort? depth = null, AttributeSelector requestedPartAttributes = null, bool withHistory = false, CancellationToken cancellationToken = default )
 		{
-			if( partUuids != null && partUuids.Length > 0 )
+			if( partUuids != null && partUuids.Count > 0 )
 			{
-				var result = new List<InspectionPlanPartDto>( partUuids.Length );
+				var result = new List<InspectionPlanPartDto>( partUuids.Count );
 				foreach( var uuid in partUuids )
 				{
 					var inspectionPlanPart = await GetPartByUuid( uuid, requestedPartAttributes, withHistory, cancellationToken ).ConfigureAwait( false );
@@ -588,19 +609,19 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 			}
 
 			var parameter = RestClientHelper.ParseToParameter( partPath, partUuids, null, depth, requestedPartAttributes, withHistory: withHistory );
-			return await _RestClient.Request<InspectionPlanPartDto[]>( RequestBuilder.CreateGet( "parts", parameter.ToArray() ), cancellationToken ).ConfigureAwait( false );
+			return await _RestClient.Request<IReadOnlyList<InspectionPlanPartDto>>( RequestBuilder.CreateGet( "parts", parameter ), cancellationToken ).ConfigureAwait( false );
 		}
 
 		/// <inheritdoc />
 		public async Task<InspectionPlanPartDto> GetPartByUuid( Guid partUuid, AttributeSelector requestedPartAttributes = null, bool withHistory = false, CancellationToken cancellationToken = default )
 		{
 			var parameter = RestClientHelper.ParseToParameter( requestedPartAttributes: requestedPartAttributes, withHistory: withHistory );
-			return await _RestClient.Request<InspectionPlanPartDto>( RequestBuilder.CreateGet( $"parts/{partUuid}", parameter.ToArray() ), cancellationToken ).ConfigureAwait( false );
+			return await _RestClient.Request<InspectionPlanPartDto>( RequestBuilder.CreateGet( $"parts/{partUuid}", parameter ), cancellationToken ).ConfigureAwait( false );
 		}
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="parts"/> is <see langword="null"/>.</exception>
-		public Task CreateParts( InspectionPlanPartDto[] parts, bool versioningEnabled = false, CancellationToken cancellationToken = default )
+		public Task CreateParts( IReadOnlyCollection<InspectionPlanPartDto> parts, bool versioningEnabled = false, CancellationToken cancellationToken = default )
 		{
 			if( parts == null ) throw new ArgumentNullException( nameof( parts ) );
 
@@ -623,7 +644,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="parts"/> is <see langword="null"/>.</exception>
-		public Task UpdateParts( InspectionPlanPartDto[] parts, bool versioningEnabled = false, CancellationToken cancellationToken = default )
+		public Task UpdateParts( IReadOnlyCollection<InspectionPlanPartDto> parts, bool versioningEnabled = false, CancellationToken cancellationToken = default )
 		{
 			if( parts == null ) throw new ArgumentNullException( nameof( parts ) );
 			var parameter = ParameterDefinition.Create( "versioningEnabled", versioningEnabled.ToString() );
@@ -632,7 +653,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="partPath"/> is <see langword="null"/>.</exception>
-		public Task DeleteParts( PathInformationDto partPath, CancellationToken cancellationToken = default )
+		public Task DeleteParts( PathInformation partPath, CancellationToken cancellationToken = default )
 		{
 			if( partPath == null ) throw new ArgumentNullException( nameof( partPath ) );
 			var parameter = ParameterDefinition.Create( "partPath", PathHelper.PathInformation2DatabaseString( partPath ) );
@@ -640,7 +661,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		}
 
 		/// <inheritdoc />
-		public async Task DeleteParts( Guid[] partUuids, CancellationToken cancellationToken = default )
+		public async Task DeleteParts( IReadOnlyCollection<Guid> partUuids, CancellationToken cancellationToken = default )
 		{
 			const string requestPath = "parts";
 			var targetSize = RestClientHelper.GetUriTargetSize( ServiceLocation, requestPath, MaxUriLength, ParameterDefinition.Create( "partUuids", "{}" ) );
@@ -672,18 +693,19 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		}
 
 		/// <inheritdoc />
-		public async Task<IEnumerable<InspectionPlanCharacteristicDto>> GetCharacteristics( PathInformationDto partPath = null, ushort? depth = null, AttributeSelector requestedCharacteristicAttributes = null, bool withHistory = false, CancellationToken cancellationToken = default )
+		public async Task<IReadOnlyList<InspectionPlanCharacteristicDto>> GetCharacteristics( PathInformation partPath = null, ushort? depth = null, AttributeSelector requestedCharacteristicAttributes = null, bool withHistory = false, CancellationToken cancellationToken = default )
 		{
 			var parameter = RestClientHelper.ParseToParameter( partPath, null, null, depth, null, requestedCharacteristicAttributes, withHistory );
-			return await _RestClient.Request<InspectionPlanCharacteristicDto[]>( RequestBuilder.CreateGet( "characteristics", parameter.ToArray() ), cancellationToken ).ConfigureAwait( false );
+			return await _RestClient.Request<IReadOnlyList<InspectionPlanCharacteristicDto>>( RequestBuilder.CreateGet( "characteristics", parameter ), cancellationToken ).ConfigureAwait( false );
 		}
 
 		/// <inheritdoc />
-		public async Task<IEnumerable<InspectionPlanCharacteristicDto>> GetCharacteristicsByUuids( Guid[] charUuids, AttributeSelector requestedCharacteristicAttributes = null, bool withHistory = false, CancellationToken cancellationToken = default )
+		public async Task<IReadOnlyList<InspectionPlanCharacteristicDto>> GetCharacteristicsByUuids( IReadOnlyCollection<Guid> charUuids, AttributeSelector requestedCharacteristicAttributes = null, bool withHistory = false, CancellationToken cancellationToken = default )
 		{
-			if( charUuids == null || charUuids.Length == 0 ) return new InspectionPlanCharacteristicDto[ 0 ];
+			if( charUuids == null || charUuids.Count == 0 )
+				return Array.Empty<InspectionPlanCharacteristicDto>();
 
-			var result = new List<InspectionPlanCharacteristicDto>( charUuids.Length );
+			var result = new List<InspectionPlanCharacteristicDto>( charUuids.Count );
 
 			var featureMatrix = await GetFeatureMatrixInternal( FetchBehavior.FetchIfNotCached, cancellationToken ).ConfigureAwait( false );
 			if( !featureMatrix.SupportsCharacteristicUuidRestrictionForCharacteristicFetch )
@@ -703,7 +725,8 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 				foreach( var uuidList in ArrayHelper.Split( charUuids, targetSize, RestClientHelper.LengthOfListElementInUri ) )
 				{
 					var parameter = RestClientHelper.ParseToParameter( null, null, uuidList, null, null, requestedCharacteristicAttributes, withHistory );
-					var characteristics = await _RestClient.Request<InspectionPlanCharacteristicDto[]>( RequestBuilder.CreateGet( "characteristics", parameter.ToArray() ), cancellationToken ).ConfigureAwait( false );
+					var characteristics = await _RestClient.Request<IReadOnlyList<InspectionPlanCharacteristicDto>>( RequestBuilder.CreateGet( "characteristics", parameter ), cancellationToken ).ConfigureAwait( false );
+
 					result.AddRange( characteristics );
 				}
 			}
@@ -715,12 +738,12 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		public async Task<InspectionPlanCharacteristicDto> GetCharacteristicByUuid( Guid charUuid, AttributeSelector requestedCharacteristicAttributes = null, bool withHistory = false, CancellationToken cancellationToken = default )
 		{
 			var parameter = RestClientHelper.ParseToParameter( requestedCharacteristicAttributes: requestedCharacteristicAttributes, withHistory: withHistory );
-			return await _RestClient.Request<InspectionPlanCharacteristicDto>( RequestBuilder.CreateGet( $"characteristics/{charUuid}", parameter.ToArray() ), cancellationToken ).ConfigureAwait( false );
+			return await _RestClient.Request<InspectionPlanCharacteristicDto>( RequestBuilder.CreateGet( $"characteristics/{charUuid}", parameter ), cancellationToken ).ConfigureAwait( false );
 		}
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="characteristics"/> is <see langword="null"/>.</exception>
-		public Task CreateCharacteristics( InspectionPlanCharacteristicDto[] characteristics, bool versioningEnabled = false, CancellationToken cancellationToken = default )
+		public Task CreateCharacteristics( IReadOnlyCollection<InspectionPlanCharacteristicDto> characteristics, bool versioningEnabled = false, CancellationToken cancellationToken = default )
 		{
 			if( characteristics == null ) throw new ArgumentNullException( nameof( characteristics ) );
 
@@ -743,7 +766,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="characteristics"/> is <see langword="null"/>.</exception>
-		public Task UpdateCharacteristics( InspectionPlanCharacteristicDto[] characteristics, bool versioningEnabled = false, CancellationToken cancellationToken = default )
+		public Task UpdateCharacteristics( IReadOnlyCollection<InspectionPlanCharacteristicDto> characteristics, bool versioningEnabled = false, CancellationToken cancellationToken = default )
 		{
 			if( characteristics == null ) throw new ArgumentNullException( nameof( characteristics ) );
 			var parameter = ParameterDefinition.Create( "versioningEnabled", versioningEnabled.ToString() );
@@ -752,7 +775,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="charPath"/> is <see langword="null"/>.</exception>
-		public Task DeleteCharacteristics( PathInformationDto charPath, CancellationToken cancellationToken = default )
+		public Task DeleteCharacteristics( PathInformation charPath, CancellationToken cancellationToken = default )
 		{
 			if( charPath == null ) throw new ArgumentNullException( nameof( charPath ) );
 			var parameter = ParameterDefinition.Create( "charPath", PathHelper.PathInformation2DatabaseString( charPath ) );
@@ -760,7 +783,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		}
 
 		/// <inheritdoc />
-		public async Task DeleteCharacteristics( Guid[] charUuid, CancellationToken cancellationToken = default )
+		public async Task DeleteCharacteristics( IReadOnlyCollection<Guid> charUuid, CancellationToken cancellationToken = default )
 		{
 			const string requestPath = "characteristics";
 			var targetSize = RestClientHelper.GetUriTargetSize( ServiceLocation, requestPath, MaxUriLength, ParameterDefinition.Create( "charUuids", "{}" ) );
@@ -773,7 +796,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		}
 
 		/// <inheritdoc />
-		public async Task<SimpleMeasurementDto[]> GetMeasurements( PathInformationDto partPath = null, MeasurementFilterAttributesDto filter = null, CancellationToken cancellationToken = default )
+		public async Task<IReadOnlyList<SimpleMeasurementDto>> GetMeasurements( PathInformation partPath = null, MeasurementFilterAttributesDto filter = null, CancellationToken cancellationToken = default )
 		{
 			await ThrowOnUnsupportedMergeAttributes( filter?.MergeAttributes, cancellationToken );
 			await ThrowOnUnsupportedMergeMasterPart( filter, cancellationToken );
@@ -782,61 +805,61 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 			const string requestPath = "measurements";
 
 			// split multiple measurement uuids into chunks of uuids using multiple requests to avoid "Request-URI Too Long" exception
-			if( filter?.MeasurementUuids?.Length > 0 )
+			if( filter?.MeasurementUuids?.Count > 0 )
 			{
 				var newFilter = filter.Clone();
 				newFilter.MeasurementUuids = null;
 
 				var parameterName = AbstractMeasurementFilterAttributesDto.MeasurementUuidsParamName;
-				var parameterDefinitions = CreateParameterDefinitions( partPath, newFilter ).ToArray();
+				var parameterDefinitions = CreateParameterDefinitions( partPath, newFilter );
 
 				//Split into multiple parameter sets to limit uuid parameter lenght
 				var parameterSets = RestClientHelper.SplitAndMergeParameters( ServiceLocation, requestPath, MaxUriLength, parameterName, filter.MeasurementUuids, parameterDefinitions );
 
 				//Execute requests in parallel
 				var requests = parameterSets
-					.Select( set => RequestBuilder.CreateGet( requestPath, set.ToArray() ) )
-					.Select( request => _RestClient.Request<SimpleMeasurementDto[]>( request, cancellationToken ) );
+					.Select( set => RequestBuilder.CreateGet( requestPath, set ) )
+					.Select( request => _RestClient.Request<IReadOnlyCollection<SimpleMeasurementDto>>( request, cancellationToken ) );
 				var result = await Task.WhenAll( requests ).ConfigureAwait( false );
 
-				return LimitAndSortResult( result.SelectMany( r => r ), filter, result.Length ).ToArray();
+				return LimitAndSortResult( result.SelectMany( r => r ), filter, result.Length );
 			}
 
 			// split multiple part uuids into chunks of uuids using multiple requests to avoid "Request-URI Too Long" exception
-			if( filter?.PartUuids?.Length > 0 )
+			if( filter?.PartUuids?.Count > 0 )
 			{
 				var newFilter = filter.Clone();
 				newFilter.PartUuids = null;
 
 				const string parameterName = AbstractMeasurementFilterAttributesDto.PartUuidsParamName;
-				var parameterDefinitions = CreateParameterDefinitions( partPath, newFilter ).ToArray();
+				var parameterDefinitions = CreateParameterDefinitions( partPath, newFilter );
 
 				//Split into multiple parameter sets to limit uuid parameter lenght
 				var parameterSets = RestClientHelper.SplitAndMergeParameters( ServiceLocation, requestPath, MaxUriLength, parameterName, filter.PartUuids, parameterDefinitions );
 
 				//Execute requests in parallel
 				var requests = parameterSets
-					.Select( set => RequestBuilder.CreateGet( requestPath, set.ToArray() ) )
-					.Select( request => _RestClient.Request<SimpleMeasurementDto[]>( request, cancellationToken ) );
+					.Select( set => RequestBuilder.CreateGet( requestPath, set ) )
+					.Select( request => _RestClient.Request<IReadOnlyList<SimpleMeasurementDto>>( request, cancellationToken ) );
 				var result = await Task.WhenAll( requests ).ConfigureAwait( false );
 
-				return LimitAndSortResult( result.SelectMany( r => r ), filter, result.Length ).ToArray();
+				return LimitAndSortResult( result.SelectMany( r => r ), filter, result.Length );
 			}
 
 			{
-				var parameterDefinitions = CreateParameterDefinitions( partPath, filter ).ToArray();
+				var parameterDefinitions = CreateParameterDefinitions( partPath, filter );
 				var requestUrl = RequestBuilder.CreateGet( requestPath, parameterDefinitions );
-				return await _RestClient.Request<SimpleMeasurementDto[]>( requestUrl, cancellationToken ).ConfigureAwait( false );
+				return await _RestClient.Request<IReadOnlyList<SimpleMeasurementDto>>( requestUrl, cancellationToken ).ConfigureAwait( false );
 			}
 		}
 
 		/// <inheritdoc />
-		public async Task<string[]> GetDistinctMeasurementAttributeValues( ushort key, PathInformationDto partPath = null, DistinctMeasurementFilterAttributesDto filter = null, CancellationToken cancellationToken = default )
+		public async Task<IReadOnlyList<string>> GetDistinctMeasurementAttributeValues( ushort key, PathInformation partPath = null, DistinctMeasurementFilterAttributesDto filter = null, CancellationToken cancellationToken = default )
 		{
 			await ThrowOnUnsupportedDistinctMeasurementValueSearch( cancellationToken );
 			await ThrowOnUnsupportedLimitResultPerPart( filter, cancellationToken );
 
-			if( filter?.MeasurementUuids?.Length > 0 )
+			if( filter?.MeasurementUuids?.Count > 0 )
 			{
 				var newFilter = filter.Clone();
 				newFilter.MeasurementUuids = null;
@@ -847,39 +870,40 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 				var requestRestriction = RequestBuilder.AppendParameters( "values", parameter );
 				var targetSize = RestClientHelper.GetUriTargetSize( ServiceLocation, requestRestriction, MaxUriLength );
 
-				var result = new List<string>( filter.MeasurementUuids.Length );
+				var result = new List<string>( filter.MeasurementUuids.Count );
+
 				foreach( var uuids in ArrayHelper.Split( filter.MeasurementUuids, targetSize, RestClientHelper.LengthOfListElementInUri ) )
 				{
 					newFilter.MeasurementUuids = uuids;
 
-					var attributes = await _RestClient.Request<string[]>( RequestBuilder.CreateGet( "distinctMeasurementAttributeValues", CreateParameterDefinitions( partPath, newFilter, key ).ToArray() ), cancellationToken ).ConfigureAwait( false );
+					var attributes = await _RestClient.Request<IReadOnlyCollection<string>>( RequestBuilder.CreateGet( "distinctMeasurementAttributeValues", CreateParameterDefinitions( partPath, newFilter, key ) ), cancellationToken ).ConfigureAwait( false );
 					result.AddRange( attributes );
 				}
 
-				return result.ToArray();
+				return result;
 			}
 
-			return await _RestClient.Request<string[]>( RequestBuilder.CreateGet( "distinctMeasurementAttributeValues", CreateParameterDefinitions( partPath, filter, key ).ToArray() ), cancellationToken ).ConfigureAwait( false );
+			return await _RestClient.Request<IReadOnlyList<string>>( RequestBuilder.CreateGet( "distinctMeasurementAttributeValues", CreateParameterDefinitions( partPath, filter, key ) ), cancellationToken ).ConfigureAwait( false );
 		}
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="measurements"/> is <see langword="null"/>.</exception>
-		public Task CreateMeasurements( SimpleMeasurementDto[] measurements, CancellationToken cancellationToken = default )
+		public Task CreateMeasurements( IReadOnlyCollection<SimpleMeasurementDto> measurements, CancellationToken cancellationToken = default )
 		{
 			if( measurements == null ) throw new ArgumentNullException( nameof( measurements ) );
-			return _RestClient.Request<SimpleMeasurementDto[]>( RequestBuilder.CreatePost( "measurements", Payload.Create( measurements ) ), cancellationToken );
+			return _RestClient.Request( RequestBuilder.CreatePost( "measurements", Payload.Create( measurements ) ), cancellationToken );
 		}
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="measurements"/> is <see langword="null"/>.</exception>
-		public Task UpdateMeasurements( SimpleMeasurementDto[] measurements, CancellationToken cancellationToken = default )
+		public Task UpdateMeasurements( IReadOnlyCollection<SimpleMeasurementDto> measurements, CancellationToken cancellationToken = default )
 		{
 			if( measurements == null ) throw new ArgumentNullException( nameof( measurements ) );
-			return _RestClient.Request<SimpleMeasurementDto[]>( RequestBuilder.CreatePut( "measurements", Payload.Create( measurements ) ), cancellationToken );
+			return _RestClient.Request( RequestBuilder.CreatePut( "measurements", Payload.Create( measurements ) ), cancellationToken );
 		}
 
 		/// <inheritdoc />
-		public async Task DeleteMeasurementsByPartPath( PathInformationDto partPath = null, GenericSearchConditionDto filter = null, AggregationMeasurementSelectionDto aggregation = AggregationMeasurementSelectionDto.Default, MeasurementDeleteBehaviorDto deep = MeasurementDeleteBehaviorDto.DeleteForCurrentPartOnly, CancellationToken cancellationToken = default )
+		public async Task DeleteMeasurementsByPartPath( PathInformation partPath = null, GenericSearchConditionDto filter = null, AggregationMeasurementSelectionDto aggregation = AggregationMeasurementSelectionDto.Default, MeasurementDeleteBehaviorDto deep = MeasurementDeleteBehaviorDto.DeleteForCurrentPartOnly, CancellationToken cancellationToken = default )
 		{
 			var parameter = new List<ParameterDefinition>();
 
@@ -903,7 +927,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		}
 
 		/// <inheritdoc />
-		public async Task DeleteMeasurementsByPartUuids( Guid[] partUuids, GenericSearchConditionDto filter = null, AggregationMeasurementSelectionDto aggregation = AggregationMeasurementSelectionDto.Default, CancellationToken cancellationToken = default )
+		public async Task DeleteMeasurementsByPartUuids( IReadOnlyCollection<Guid> partUuids, GenericSearchConditionDto filter = null, AggregationMeasurementSelectionDto aggregation = AggregationMeasurementSelectionDto.Default, CancellationToken cancellationToken = default )
 		{
 			if( partUuids == null )
 			{
@@ -935,7 +959,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		}
 
 		/// <inheritdoc />
-		public async Task DeleteMeasurementsByUuid( Guid[] measurementUuids, CancellationToken cancellationToken = default )
+		public async Task DeleteMeasurementsByUuid( IReadOnlyCollection<Guid> measurementUuids, CancellationToken cancellationToken = default )
 		{
 			if( measurementUuids.Any() )
 			{
@@ -953,23 +977,23 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 		}
 
 		/// <inheritdoc />
-		public async Task<DataMeasurementDto[]> GetMeasurementValues( PathInformationDto partPath = null, MeasurementValueFilterAttributesDto filter = null, CancellationToken cancellationToken = default )
+		public async Task<IReadOnlyList<DataMeasurementDto>> GetMeasurementValues( PathInformation partPath = null, MeasurementValueFilterAttributesDto filter = null, CancellationToken cancellationToken = default )
 		{
 			await ThrowOnUnsupportedMergeAttributes( filter?.MergeAttributes, cancellationToken );
 			await ThrowOnUnsupportedLimitResultPerPart( filter, cancellationToken );
 
-			if( filter?.MeasurementUuids?.Length > 0 )
+			if( filter?.MeasurementUuids?.Count > 0 )
 				return await GetMeasurementValuesSplitByMeasurement( partPath, filter, cancellationToken ).ConfigureAwait( false );
 
-			if( filter?.CharacteristicsUuidList?.Length > 0 )
+			if( filter?.CharacteristicsUuidList?.Count > 0 )
 				return await GetMeasurementValuesSplitByCharacteristics( partPath, filter, cancellationToken ).ConfigureAwait( false );
 
-			return await _RestClient.Request<DataMeasurementDto[]>( RequestBuilder.CreateGet( "values", CreateParameterDefinitions( partPath, filter ).ToArray() ), cancellationToken ).ConfigureAwait( false );
+			return await _RestClient.Request<IReadOnlyList<DataMeasurementDto>>( RequestBuilder.CreateGet( "values", CreateParameterDefinitions( partPath, filter ) ), cancellationToken ).ConfigureAwait( false );
 		}
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="values"/> is <see langword="null"/>.</exception>
-		public Task CreateMeasurementValues( DataMeasurementDto[] values, CancellationToken cancellationToken = default )
+		public Task CreateMeasurementValues( IReadOnlyCollection<DataMeasurementDto> values, CancellationToken cancellationToken = default )
 		{
 			if( values == null ) throw new ArgumentNullException( nameof( values ) );
 			return _RestClient.Request( RequestBuilder.CreatePost( "values", Payload.Create( values ) ), cancellationToken );
@@ -977,7 +1001,7 @@ namespace Zeiss.PiWeb.Api.Rest.HttpClient.Data
 
 		/// <inheritdoc />
 		/// <exception cref="ArgumentNullException"><paramref name="values"/> is <see langword="null"/>.</exception>
-		public Task UpdateMeasurementValues( DataMeasurementDto[] values, CancellationToken cancellationToken = default )
+		public Task UpdateMeasurementValues( IReadOnlyCollection<DataMeasurementDto> values, CancellationToken cancellationToken = default )
 		{
 			if( values == null ) throw new ArgumentNullException( nameof( values ) );
 			return _RestClient.Request( RequestBuilder.CreatePut( "values", Payload.Create( values ) ), cancellationToken );
